Rename `GenModel` to `GenDBModel`
